requirements.txt
Django==1.6.5
PIL==1.1.7
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
django-bower==5.0.1
django-celery==3.1.10
django-crispy-forms==1.4.0
django-dropbox==0.0.2
django-social-auth==0.7.28
dropbox==2.1.0
httplib2==0.9
kombu==3.0.21
numpy==1.8.1
oauth2==1.5.211
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
six==1.7.3
urllib3==1.9
wsgiref==0.1.2
